move newsroom to the separate components
