+++ css/styles.css
@@ -2,6 +2,7 @@
 @import "tailwindcss/base";
 @import "tailwindcss/components";
 @import url(https://cdn.jsdelivr.net/gh/tonsky/FiraCode@4/distr/fira_code.css);
+@import "react-responsive-carousel/lib/styles/carousel.min.css";
 /*! purgecss end ignore */
 
 @import "tailwindcss/utilities";
