+++ package.json
@@ -61,6 +61,7 @@
     "react-dom": "^17.0.2",
     "react-ga": "^3.1.2",
     "react-gtm-module": "^2.0.11",
+    "react-responsive-carousel": "^3.2.23",
     "react-scrollspy": "^3.4.2",
     "react-syntax-highlighter": "^12.2.1",
     "react-text-truncate": "^0.16.0",
